Add files via upload
These are two variations of the WS2812b LED lighting control program, utilizing the ADXL345 accelerometer sensor as input. 
"X_Hue_Y_Sat_Z_Light.ino" Uses the X, Y, and Z axes to control the H, S, and L 'sliders' in 3 dimensions. 
"XY_Adjust_Hue_Fixed_SL.ino" Is similar but it uses the X and Y axes to adjust the hue, and it fixes the saturation and lightness to max values. This allows for more vibrant colors that are more consistent. Just a slightly different efect.